p4i integration Integrate 275576 from 4.3 to main: Fixes:    Doc: Changed undo framework example to work on gcc 3.3 Task:     155910
[git-p4: depot-paths = "//depot/qt/main/": change = 275577]

+++ examples/tools/undoframework/diagramitem.cpp
@@ -29,7 +29,8 @@ DiagramItem::DiagramItem(DiagramType diagramType, QGraphicsItem *item,
         setPolygon(trianglePolygon);
     }
 
-    QColor color(qrand() % 256, qrand() % 256, qrand() % 256);
+    QColor color(static_cast<int>(qrand()) % 256,
+        static_cast<int>(qrand()) % 256, static_cast<int>(qrand()) % 256);
     QBrush brush(color);
     setBrush(brush);
     setFlag(QGraphicsItem::ItemIsSelectable);
